requirements.txt
asyncpg
bcrypt==4.0.1
fastapi
langchain
langchain-community
langchain_google_genai
mypy
openai==0.28
passlib[bcrypt]
pre-commit
psycopg2-binary
pydantic
pydantic[email]
pylint
python-dotenv
python-jose
python-multipart
sqlalchemy
uvicorn
